ProductDetail_page
in Home page where all the product list were shown if user clicks on any of the product img or price they will redirect on the productDetail page where they will get the full detail about the product like price and its reviews. this all would happened because we are using the dynamic route feature.